final submit

+++ src/Controller/CategorieController.php
@@ -4,7 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Categorie;
 use App\Form\CategorieFormType;
-use App\Repository\ClientRepository;
+use App\Repository\ProduitRepository;
 use App\Repository\CategorieRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\ManagerRegistry;
@@ -12,14 +12,16 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use App\Entity\Client;
 
 class CategorieController extends AbstractController
 {
     protected CategorieRepository $CategorieRepository;
+    protected ProduitRepository $produitRepository;
     public function __construct(
+        ProduitRepository $produitRepository,
         CategorieRepository $CategorieRepository,
     ) {
+        $this->produitRepository = $produitRepository;
         $this->CategorieRepository = $CategorieRepository;
     }
 
@@ -71,7 +73,16 @@ class CategorieController extends AbstractController
         $routeParams = $request->attributes->get('_route_params');
         $id = $routeParams['id'];
         $categorie = $this->CategorieRepository->find($id);
-
+        $produits = $this->produitRepository->findBy(array("Categorie" => $categorie));
+        if ($produits) {
+            $categories = $this->CategorieRepository->findAll();
+            $error = "oui";
+            return $this->render('categorie/index.html.twig', [
+                'controller_name' => 'CategorieController',
+                'categories' => $categories,
+                'error' => $error
+            ]);
+        }
         $entityManager = $doctrine->getManager();
         $entityManager->remove($categorie);
         $entityManager->flush();
